+++ README.ru.md
@@ -98,6 +98,7 @@ docker compose up -d
 
 ## Contributors
 
+* [Anton Glushakov](https://github.com/glushakov)
 * [Mikhail Grigorev](https://github.com/cherts)
 * [Ilya Lukyanov](mailto:lukyanov1985@gmail.com)
 * [Roman Minebaev](https://github.com/minebaev)
