+++ frontend/src/locales/ru_RU.json
@@ -619,10 +619,7 @@
     "hostRole": {
         "primary": "Мастер",
         "replica": "Реплика",
-        "unknown": "Роль неизвестна",
-        "primaryHint": "Хост принимает запись",
-        "replicaHint": "Хост в режиме восстановления: только чтение",
-        "unknownHint": "Не удалось определить роль хоста"
+        "unknown": "Роль неизвестна"
     },
     "Common summary": "Основная информация",
     "home": {
